+++ maya_viewport_match/metadata.py
@@ -1,2 +1,2 @@
-VERSION = (0, 30, 4)
+VERSION = (0, 34, 1)
 VERSION_STRING = ".".join(str(part) for part in VERSION)